backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
psycopg2-binary
alembic
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-docx
Pillow
PyPDF2
pdfplumber
selenium
webdriver-manager
requests
beautifulsoup4
redis
celery
pytesseract
opencv-python-headless
openai
python-magic
python-magic-bin; platform_system == "Windows"
aiofiles
httpx
numpy
Jinja2
PyMuPDF
lxml
scikit-learn
scipy
pandas
nltk
spacy
transformers
torch
docling
langchain
langchain-openai
langchain-community
langchain-core
pydantic-settings
python-dotenv
aiohttp
asyncio-throttle
easyocr
pytz 